Added inline styling to both IraqiKurdistanRegion and KETA
